requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
scikit-learn==1.3.2
mplsoccer==1.1.12
tabulate==0.9.0
python-dotenv==1.0.0
streamlit==1.32.0
scipy==1.11.4
